Basic outline before getting screen working.
Neopixel on mech key now fades and changes color along with getting the RTC code ready. Code right now does next to nothing on the beginner kit but has a good portion of the basic structure. Hopefully I'll regain sanity and use the extreme amounts of spare time I have tomorrow to get the display to show a hello world or something similar.