Screen now showing a hello world and cleaned up code.
Cleaned up code, changed the "owo" var to the sensible name of color, display is working but is upside down (trying to fix, library is confusing) also either the library is very inefficient or my code is causing a massive memory kerfuffle as with just a hello world 77% of the memory is used. This memory issue could be left in without going over the limit by Friday but in later revisions it should be resolved.